+++ src/Modules/Unit/DTO/Unit.php
@@ -12,35 +12,28 @@ class Unit
 	/**
 	 * Identifer of the unit.
 	 */
-	public int $id;
+	public int $id = 0;
 
 	/**
 	 * Label for a unit.
 	 */
-	public string $name;
+	public string $name = '';
 
 	/**
 	 * @see UnitType
 	 */
-	public ?int $type;
-
-	public function __construct()
-	{
-		$this->id = 0;
-		$this->name = '';
-		$this->type = UnitType::UNDEFINED;
-	}
+	public ?int $type = UnitType::UNDEFINED;
 
 	/**
 	 * Creates a unit out of an array representation like the Database select.
 	 */
-	public static function createFromArray($query_result, $prefix = ''): Unit
+	public static function createFromArray($queryResult, $prefix = ''): Unit
 	{
-		$obj = new Unit();
-		$obj->id = $query_result["{$prefix}id"];
-		$obj->name = $query_result["{$prefix}name"];
-		$obj->type = $query_result["{$prefix}type"];
+		$unit = new Unit();
+		$unit->id = $queryResult["{$prefix}id"];
+		$unit->name = $queryResult["{$prefix}name"];
+		$unit->type = $queryResult["{$prefix}type"];
 
-		return $obj;
+		return $unit;
 	}
 }