+++ src/RestApi/UserRestController.php
@@ -17,6 +17,7 @@ use Foodsharing\Modules\Region\RegionTransactions;
 use Foodsharing\Modules\Register\DTO\RegisterData;
 use Foodsharing\Modules\Register\RegisterTransactions;
 use Foodsharing\Modules\Settings\SettingsGateway;
+use Foodsharing\Modules\Unit\DTO\UserUnit;
 use Foodsharing\Modules\Uploads\UploadsGateway;
 use Foodsharing\Permissions\BlogPermissions;
 use Foodsharing\Permissions\ContentPermissions;
@@ -46,90 +47,38 @@ use Symfony\Component\HttpKernel\Exception\UnauthorizedHttpException;
 
 class UserRestController extends AbstractFOSRestController
 {
-	private Session $session;
-	private LoginGateway $loginGateway;
-	private FoodsaverGateway $foodsaverGateway;
-	private ProfileGateway $profileGateway;
-	private UploadsGateway $uploadsGateway;
-	private RegionGateway $regionGateway;
-	private SettingsGateway $settingsGateway;
-	private RegionTransactions $regionTransactions;
-	private GroupTransactions $groupTransactions;
-
-	private UserPermissions $userPermissions;
-	private ProfilePermissions $profilePermissions;
-	private BlogPermissions $blogPermissions;
-	private ContentPermissions $contentPermissions;
-	private MailboxPermissions $mailboxPermissions;
-	private NewsletterEmailPermissions $newsletterEmailPermissions;
-	private QuizPermissions $quizPermissions;
-	private RegionPermissions $regionPermissions;
-	private ReportPermissions $reportPermissions;
-	private StorePermissions $storePermissions;
-	private WorkGroupPermissions $workGroupPermissions;
-
-	private EmailHelper $emailHelper;
-	private RegisterTransactions $registerTransactions;
-	private ProfileTransactions $profileTransactions;
-	private FoodsaverTransactions $foodsaverTransactions;
-
 	private const MIN_RATING_MESSAGE_LENGTH = 100;
 	private const MIN_PASSWORD_LENGTH = 8;
 	private const MIN_AGE_YEARS = 18;
 	private const DELETE_USER_MAX_REASON_LEN = 200;
 
 	public function __construct(
-		Session $session,
-		LoginGateway $loginGateway,
-		FoodsaverGateway $foodsaverGateway,
-		ProfileGateway $profileGateway,
-		UploadsGateway $uploadsGateway,
-		RegionGateway $regionGateway,
-		EmailHelper $emailHelper,
-		RegisterTransactions $registerTransactions,
-		ProfileTransactions $profileTransactions,
-		FoodsaverTransactions $foodsaverTransactions,
-		SettingsGateway $settingsGateway,
+		private Session $session,
+		private LoginGateway $loginGateway,
+		private FoodsaverGateway $foodsaverGateway,
+		private ProfileGateway $profileGateway,
+		private UploadsGateway $uploadsGateway,
+		private RegionGateway $regionGateway,
+		private EmailHelper $emailHelper,
+		private RegisterTransactions $registerTransactions,
+		private ProfileTransactions $profileTransactions,
+		private FoodsaverTransactions $foodsaverTransactions,
+		private SettingsGateway $settingsGateway,
 
-		UserPermissions $userPermissions,
-		ProfilePermissions $profilePermissions,
-		MailboxPermissions $mailboxPermissions,
-		QuizPermissions $quizPermissions,
-		ReportPermissions $reportPermissions,
-		StorePermissions $storePermissions,
-		ContentPermissions $contentPermissions,
-		BlogPermissions $blogPermissions,
-		RegionPermissions $regionPermissions,
-		NewsletterEmailPermissions $newsletterEmailPermissions,
-		WorkGroupPermissions $workGroupPermissions,
-		RegionTransactions $regionTransactions,
-		GroupTransactions $groupTransactions
+		private UserPermissions $userPermissions,
+		private ProfilePermissions $profilePermissions,
+		private MailboxPermissions $mailboxPermissions,
+		private QuizPermissions $quizPermissions,
+		private ReportPermissions $reportPermissions,
+		private StorePermissions $storePermissions,
+		private ContentPermissions $contentPermissions,
+		private BlogPermissions $blogPermissions,
+		private RegionPermissions $regionPermissions,
+		private NewsletterEmailPermissions $newsletterEmailPermissions,
+		private WorkGroupPermissions $workGroupPermissions,
+		private RegionTransactions $regionTransactions,
+		private GroupTransactions $groupTransactions
 	) {
-		$this->session = $session;
-		$this->loginGateway = $loginGateway;
-		$this->foodsaverGateway = $foodsaverGateway;
-		$this->profileGateway = $profileGateway;
-		$this->uploadsGateway = $uploadsGateway;
-		$this->regionGateway = $regionGateway;
-		$this->emailHelper = $emailHelper;
-		$this->registerTransactions = $registerTransactions;
-		$this->profileTransactions = $profileTransactions;
-		$this->foodsaverTransactions = $foodsaverTransactions;
-		$this->settingsGateway = $settingsGateway;
-		$this->regionTransactions = $regionTransactions;
-		$this->groupTransactions = $groupTransactions;
-
-		$this->userPermissions = $userPermissions;
-		$this->profilePermissions = $profilePermissions;
-		$this->mailboxPermissions = $mailboxPermissions;
-		$this->quizPermissions = $quizPermissions;
-		$this->reportPermissions = $reportPermissions;
-		$this->storePermissions = $storePermissions;
-		$this->contentPermissions = $contentPermissions;
-		$this->blogPermissions = $blogPermissions;
-		$this->regionPermissions = $regionPermissions;
-		$this->newsletterEmailPermissions = $newsletterEmailPermissions;
-		$this->workGroupPermissions = $workGroupPermissions;
 	}
 
 	/**
@@ -237,20 +186,11 @@ class UserRestController extends AbstractFOSRestController
 
 			// load region
 			$regions = $this->regionTransactions->getUserRegions($this->session->id());
-
-			$rsp_regions = [];
-			foreach ($regions as $region) {
-				$rsp_regions[] = UserRegionModel::createFrom($region);
-			}
-			$response['regions'] = $rsp_regions;
+			$response['regions'] = array_map(fn (UserUnit $region): UserRegionModel => UserRegionModel::createFrom($region), $regions);
 
 			// load groups
 			$groups = $this->groupTransactions->getUserGroups($this->session->id());
-			$rsp_groups = [];
-			foreach ($groups as $group) {
-				$rsp_groups[] = UserGroupModel::createFrom($group);
-			}
-			$response['groups'] = $rsp_groups;
+			$response['groups'] = array_map(fn (UserUnit $group): UserGroupModel => UserGroupModel::createFrom($group), $groups);
 		}
 
 		if ($mayAdministrateUserProfile) {